backend/requirements.txt
# 核心框架
fastapi>=0.104.1
uvicorn>=0.24.0.post1
langchain>=0.1.0
langchain-community>=0.0.19
langchain-chroma>=0.1.1
chromadb>=0.4.22

# 智谱AI SDK
zhipuai>=2.0.0

# 文件解析
python-docx>=1.1.0
PyPDF2>=3.0.1

# 环境变量/工具
python-dotenv>=1.0.0
requests>=2.31.0